+++ packages/cli/src/workspace-packages.d.ts
@@ -16,6 +16,7 @@ declare module "@zachariahredfield/playbook-engine" {
   export const loadConfig: (...args: any[]) => Promise<any>;
   export const generateRepositoryHealth: (...args: any[]) => any;
   export const generateRepositoryIndex: (...args: any[]) => any;
+  export const loadAiContract: (...args: any[]) => any;
 
   export const queryRepositoryIndex: (...args: any[]) => any;
   export const queryDependencies: (...args: any[]) => any;